First stable version of Serverus
Finished the first stable alpha version of Serverus. This version
includes the following features.

RegisterServer Function

This function receives a grpc Server that is injected by the main
application that integrates serverus.

Pendings implementations

We have pending the UnaryServerInterceptor that should allow to inject
many interceptor for a unary interceptor.

Integrate StreamServerInterceptor so we can support the stream
connection.